Fixed unit tests
Unit tests were expecting a precise filesize for the converted file, which
is a problem as it will depend of the ffmpeg installation and other factors.

I changed them so that they expect an range, but that's not ideal neither and
certainly not a good way to check that the conversion succeeded.

Ideally the resulting files should be opened with some Java MP3 library and
checked for sanity.