Improved DB generator.
Improved DB retro-generator by retrieving the table name from the SPECIFICATION
field from the class, instead of converting the class name in a SQL table name
as some tables names are slightly different from the class name.

Dealt with some fields not following the naming convention (lower case,
underscore and numbers). Some names were in camel case in the DB.

Re-ordered the field in some classes so that it matches the order in the DB,
taken from the DB documentation.

Added comment on datamodel classes whenever relevant (comment taken from the
DB documentation, as is. Accuracy not guaranteed !)